git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/summary-reuse@43400 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/predicates/pathformula/pointeraliasing/PointerTargetPattern.java
@@ -47,7 +47,7 @@ class PointerTargetPattern implements Serializable, Predicate<PointerTarget> {
     return new PointerTargetPattern(base, null, 0L, 0L);
   }
 
-  static Predicate<PointerTarget> forRange(String base, long startOffset, int size) {
+  static Predicate<PointerTarget> forRange(String base, long startOffset, long size) {
     return new RangePointerTargetPattern(base, startOffset, size);
   }
 
@@ -62,7 +62,7 @@ class PointerTargetPattern implements Serializable, Predicate<PointerTarget> {
     return lhs.accept(v).build();
   }
 
-  Predicate<PointerTarget> withRange(final int size) {
+  Predicate<PointerTarget> withRange(final long size) {
     assert containerOffset != null && properOffset != null : "Starting address is inexact";
     return forRange(base, containerOffset + properOffset, size);
   }
@@ -122,7 +122,7 @@ class PointerTargetPattern implements Serializable, Predicate<PointerTarget> {
     private final long endOffset;
 
     private RangePointerTargetPattern(
-        final String pBase, final long pStartOffset, final int pSize) {
+        final String pBase, final long pStartOffset, final long pSize) {
       base = pBase;
       startOffset = pStartOffset;
       endOffset = pStartOffset + pSize;
